// app_web/client-sn/interfaces/DatoClinico.ts
// Generated by https://quicktype.io

export interface DatoClinico {
    id: number;
    rela_users: number;
    presion_alta: string;
    presion_baja: string;
    pulso: string;
    peso: string;
    circunferencia_cintura: string;
    talla: string;
    consume_alcohol: number;
    consume_marihuana: number;
    otras_drogas: number;
    fuma_tabaco: number;
    estado_clinico: number;
    user: User;
    fecha_alta: string
}

export interface User {
    id: number;
    email: string;
    rela_rol: number;
}